Test abm (#5)
I've refactored the code for the A_bm function for a few reasons:

- The `return` statements on rows 83 and 88 of the original version will
cause the function to return after the first iteration of the loop i.e.
when `i = 0`, without running any other iterations. This means that a
single value will always be returned, no matter how many elements in `D`
and `z`.
- Numpy provides an
[`np.where()`](https://numpy.org/doc/stable/reference/generated/numpy.where.html)
function that acts like an `if-else` statement. It selects an element
from one of two arrays depending on whether a condition is met or not.
- Both array options are outputs from other calculations so I've made
those calculations into stand-alone helper functions. This makes it
easier to test both functions are working, then test that the
`np.where()` function is working correctly. Note that a preceding `_` is
used by convention on a function names to indicate that they're private
and it shouldn't be called from another module. That should really apply
to almost all functions in this module, since most of them should only
be called from inside this module, but I haven't bothered adding the `_`
anywhere else.

The downside to this approach is that you have to calculate the full
array for both options, even though only half of the total elements
across both arrays are needed after the `np.where()`. I don't think this
matters much, these functions should be quick to run. We can profile the
code later an pinpoint the slow functions and see if this is causing a
bottleneck.

The code is still roughly the same number of lines, but I think this is
a bit clearer / more readable than the way you'd written it originally.

I've also added a Mathcad sheet with the calculations I used as expected
unit test values. I wrote this independently, straight from the code so
it's really an independent check of what's in the Python module.

There's also a few other formatting changes. I use
[`black`](https://pypi.org/project/black/) to format the modules
automatically so I didn't have to do all the spacing by hand.

---------

Co-authored-by: Stuart Roy <stuart.roy@genesisoilandgas.com>