Test Q_v (#12)
Another interesting one.....

There's a difference between
[`np.min()`](https://numpy.org/doc/stable/reference/generated/numpy.min.html)
and
[`np.minimum()`](https://numpy.org/doc/stable/reference/generated/numpy.minimum.html).
`np.min()` finds the minimum value in an array - that is, it returns a
single value regardless of the array size. `np.minimum()` returns the
element minimums between 2 arrays - that is, given 2 arrays of matching
size it will return a 3rd array of the same size containing the minimum
of both arrays for a given index.

Your original implementation used `np.min()`, which worked fine when I
tested it with single values because there was only 1 minimum to check.
When I tried to pass in arrays of inputs I got some wrong answers
because I was using the same (minimum) value for all calculations.

Another subtle bug that would have been really hard to catch without
unit tests!

Also worth noting is that Python doesn't distinguish between `*` and `/`
for precedence in calculations (which makes sense since one is the
reciprocal of the other). When you mix both symbols in an expression
they are evaluated from left to right. This means your original code
using `Abm(D, z) / D * S_u)` (original line 48) is effectively
interpreted as `(Abm(D, z) / D) * S_u)` because the division is done
then the multiplication. This gave the wrong answer so I changed that
line to read `Abm(D, z) / D / S_u` (replacement line 73) and my test
passed.

Lastly, you asked a while ago why the standard deviation of shear
strength had a gradient (on Teams, 8th May). I wondered when I reviewed
this function if it's to calculate the shear strength at a depth of
penetration. The nomenclature in F114 notes that `S_u` is taken at the
pipe invert and is therefor a function of z. I haven't really thought
about it much (a problem for later) but I suppose you could draw a
random shear strength at the mudline from the distribution and a random
gradient from the distribution and then you can calculate the shear
strength at any depth for that combination. I'm not sure it both values
should be taken as independent of each other. Probably a question for
Mustafa.....

---------

Co-authored-by: Stuart Roy <stuart.roy@genesisoilandgas.com>