Adds test for k_lay function (#14)
Adds unit test for `k_lay()` function. The main change is to pass `EI`
into the function as a parameter, rather than defining it as a global
variable at the top of the module. This is useful for a few reasons:
1. Global variables can be changed by an other function, so it's easy to
accidentally change it at some point in the code an lose track of the
value stored in the variable.
2. You don't really know how you want to define the input data -
although I'd suggest that a `csv` file is one of the worst options. Have
a look at more flexible / readable formats like [JSON
](https://www.w3schools.com/whatis/whatis_json.asp)and
[TOML](https://toml.io/en/).
3. It was hard to unit test (often a sign of poor code structure)
because I would have to change the value in the CSV to test a different
value. This isn't easy to automate.

I also added a few TODOs:

- however the su gradient is calculated needs to be implemented in
there.
- `I` should be a function of `OD` and `wt` and should be calculated
from those distributions, so you can't have a single `EI` value.
- If `I` is a distribution we'd have to pass in an separate
`E`....should this also be a distribution?

Co-authored-by: Stuart Roy <stuart.roy@genesisoilandgas.com>